Создаст отдельные модули и подключит их в main
