+++ index.html
@@ -8,7 +8,7 @@
     <link rel="stylesheet" href="css/style.css">
     <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
     <script src="./js/functions.js" async></script>
-    <script src="./js/main.js" async></script>
+    <script src="./js/main.js" type="module" async></script>
     <title>Кекстаграм</title>
 </head>
